requirements.txt
crytic_compile==0.2.2
slither_analyzer==0.8.2
solc_select==0.2.1
networkx==2.5.1
z3-solver==4.11.2.0
